Feature/ai walls

+++ bobberick-framework/src/entity/components/TransformComponent.h
@@ -10,7 +10,7 @@ class TransformComponent : public Component
 public:
 	TransformComponent();
 	explicit TransformComponent(int sc);
-	explicit TransformComponent(float xPos, float yPos, int h, int w, int sc);
+	explicit TransformComponent(float xPos, float yPos, int h, int w, double sc);
 
     Vector2D position = Vector2D(0, 0);
     Vector2D velocity = Vector2D(0, 0);
@@ -25,11 +25,11 @@ public:
     void update() override;
     void init() override;
 
-    int getScale() const;
-    void setScale(const int nScale);
+    double getScale() const;
+    void setScale(const double nScale);
 
 protected:
-	int scale;
+	double scale;
 };
